Add bundled test run to ci. Remove teardown test.
The init/teardown test acts flaky, crashing the test harness from time
to time. I've never seen a crash like that in production from the
json-query crate, so I don't know what the deal is.

Not going to make the suite flaky in support of a useless usecase (init
then immediately disposing of the jq engine).